Computed Columns!
Consisted of:
Removed using namespace std and rewrote a bunch of stuff

Columns can now be overwritten from JASP-Engine/R

Rewrote a bunch of stuff

computeColumn replies added

started on the interface

Columns can be added and depend on each other

 some more changes

Refactored datasetpackage

Fixed periodic crash

refactoring and saving/loading computed columns

Refactored option a little bit and added define for engine messages

Analyses are updated nicely and columns can be removed!

Can now sync and stuff

Finishing touches

Small fixes and documentation filled

+++ JASP-Desktop/importers/importdataset.cpp
@@ -17,12 +17,12 @@ void ImportDataSet::addColumn(ImportColumn *column)
 	_columns.push_back(column);
 }
 
-int ImportDataSet::columnCount() const
+size_t ImportDataSet::columnCount() const
 {
 	return _columns.size();
 }
 
-int ImportDataSet::rowCount() const
+size_t ImportDataSet::rowCount() const
 {
 	if (columnCount() == 0)
 		return 0;
